REFACTOR(All): Create folder services
