+++ 0_Game_System/src/services/database/dbConnection.ts
@@ -1,4 +1,4 @@
-import supabase, {bucket} from '@/components/services/database/supabase'
+import supabase, {bucket} from '@/services/database/supabase'
 
 const dbConnection = supabase
 export const bucketName = bucket
