+++ 0_Game_System/tsconfig.json
@@ -27,7 +27,8 @@
       "@components*": ["./src/components/*"],
       "@pages*": ["./src/components/pages/*"],
       "@interfaces*": ["./src/components/interfaces/*"],
-      "@database*": ["./src/components/services/database/*"],
+      "@services*": ["./src/services/*"],
+      "@database*": ["./src/services/database/*"],
       "@UI*": ["./src/components/UI/*"],
       "@Icons*": ["./src/components/UI/Icons/*"],
       "@utils*": ["./src/components/utils/*"],
