+++ 0_Game_System/vite.config.ts
@@ -21,7 +21,8 @@ export default defineConfig({
       "@UI": `${path.resolve(__dirname, "./src/components/UI/")}`,
       "@Icons": `${path.resolve(__dirname, "./src/components/UI/Icons/")}`,
       "@pages": `${path.resolve(__dirname, "./src/components/pages/")}`,
-      "@database": `${path.resolve(__dirname, "./src/components/database/")}`,
+      "@services": `${path.resolve(__dirname, "./src/services/")}`,
+      "@database": `${path.resolve(__dirname, "./src/services/database/")}`,
       "@interfaces": `${path.resolve(__dirname, "./src/components/interfaces/")}`,
       "@utils": `${path.resolve(__dirname, "./src/components/utils/")}`,
     },
